pyproject.toml
[project]
name = "notes-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.117.1",
    "google-genai>=1.45.0",
    "pillow>=12.0.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "uvicorn>=0.37.0",
]
